requirements.txt
# Core Stuff
# -------------------------------------
Django==3.0.8

# Configuration
# -------------------------------------
django-environ==0.4.5

# REST APIs
# -------------------------------------
djangorestframework==3.11.0

# Models
# -------------------------------------
psycopg2-binary==2.8.6
django-uuid-upload-path==1.0.0
django-versatileimagefield==2.0
django-extensions==3.0.3

# Testing
# -------------------------------------
flake8-mypy==17.8.0
pytest==6.1.2
pytest-django==3.9.0
pytest-cov==2.9.0

# Generating DB Schema
# --------------------------------------
pydot==1.4.1
pyparsing==2.4.7
